Issue #183: another work in progress commit to save the state of play while I look at something else.
Added a new concept of a construction subsequence, which makes it easier to create diagrams of the high level structure, using new GraphViz file generation scripts in the MS Access DB Editor.

With this, CP sequencing is now fixed in the Core package and packages covering Physical, User, Access Context, Network, Network Connectivity, Application, Process Comms and Data Lifecycle submodels, except for the data cache elements in the Data Lifecycle model (due to an unnecessary dependency on the Process Trust model). I plan to leave that dependency unfixed for the moment, so the next steps will be to fix the Process Trust model dependencies, then data caching and data fields, and then Virtualisation and everything built on top of that, and finally Local Network Connectivity (pairing), IoT, Privacy and everything built on top of that.