Issue #176: added new threats modelling exploitation of vulnerabilities in processes that control network routing or network access.
The idea is that if a subnet, interface or logical path segment is 'physically' accessible to an attacker - i.e., they are able to attempt to use it (even if they are not authorized) - then they can send a message that will be handled by the process that controls the network asset. The process should enforce the policy and reject or drop the message, but if the process contains a vulnerability, it could be exploited by the attacker's message. To handle this, a new TW attribute representing 'physical network accessibiity' was added, and a number of existing threats refactored to make use of it.